add economic_functions.py , move get_population,plot_population,... to it
